+++ Sketch/code/jackson-core-2.18.1.jar
