+++ Sketch/server/obstacles.csv
@@ -1,0 +1,2 @@
+id,xPos,yPos,radius
+0.0,1.0,0.0,0.1
