--- obstacles.csv
@@ -1,66 +1,0 @@
-id,xPos,yPos,radius
-0.0,1.0,0.0,0.1
-1.0,0.0,1.0,0.1
-2.0,0.0,0.0,0.1
-3.0,1.0,1.0,0.1
-4.0,0.01,0.1825,0.1
-5.0,0.01,0.1825,0.1
-6.0,0.97875,0.18125,0.1
-7.0,0.97875,0.18125,0.1
-8.0,0.97875,0.18125,0.1
-9.0,0.98125,0.3275,0.1
-10.0,0.98125,0.3275,0.1
-11.0,0.98125,0.3275,0.1
-12.0,0.98125,0.3275,0.1
-13.0,0.9825,0.495,0.1
-14.0,0.9825,0.495,0.1
-15.0,0.9825,0.495,0.1
-16.0,0.015,0.32125,0.1
-17.0,0.015,0.32125,0.1
-18.0,0.015,0.32125,0.1
-19.0,0.015,0.32125,0.1
-20.0,0.015,0.32125,0.1
-21.0,0.025,0.45625,0.1
-22.0,0.025,0.45625,0.1
-23.0,0.025,0.45625,0.1
-24.0,0.025,0.45625,0.1
-25.0,0.02625,0.5875,0.1
-26.0,0.02625,0.5875,0.1
-27.0,0.02625,0.5875,0.1
-28.0,0.03,0.72125,0.1
-29.0,0.03,0.72125,0.1
-30.0,0.03,0.72125,0.1
-31.0,0.04125,0.8675,0.1
-32.0,0.04125,0.8675,0.1
-33.0,0.04125,0.8675,0.1
-34.0,0.9625,0.675,0.1
-35.0,0.9625,0.675,0.1
-36.0,0.9625,0.675,0.1
-37.0,0.9625,0.675,0.1
-38.0,0.955,0.82,0.1
-39.0,0.955,0.82,0.1
-40.0,0.955,0.82,0.1
-41.0,0.955,0.82,0.1
-42.0,0.1425,0.03,0.1
-43.0,0.1425,0.03,0.1
-44.0,0.1425,0.03,0.1
-45.0,0.26125,0.02875,0.1
-46.0,0.26125,0.02875,0.1
-47.0,0.26125,0.02875,0.1
-48.0,0.26125,0.02875,0.1
-49.0,0.40125,0.0275,0.1
-50.0,0.40125,0.0275,0.1
-51.0,0.40125,0.0275,0.1
-52.0,0.4025,0.0275,0.1
-53.0,0.5275,0.03625,0.1
-54.0,0.5275,0.03625,0.1
-55.0,0.5275,0.03625,0.1
-56.0,0.6725,0.025,0.1
-57.0,0.6725,0.025,0.1
-58.0,0.6725,0.025,0.1
-59.0,0.825,0.0375,0.1
-60.0,0.825,0.0375,0.1
-61.0,0.825,0.0375,0.1
-62.0,0.91125,0.06,0.1
-63.0,0.91125,0.06,0.1
-64.0,0.91125,0.06,0.1
